BMI Calculator update is done. Next step is write testcase

+++ src/main/java/com/bmicheck/service/BMIService.java
@@ -3,5 +3,7 @@ package com.bmicheck.service;
 import com.bmicheck.entity.BMIEntity;
 
 public interface BMIService {
-    public BMIEntity calculateBMI(Integer height, Integer weight);
+    public BMIEntity calculateBMI(Float height, Float weight);
+
+
 }
